requirements.txt
flask==2.0.1
flask-cors==3.0.10
google-generativeai==0.3.1
Pillow==9.0.1
python-dotenv==0.19.2
requests==2.27.1
gunicorn==20.1.0